Fix record invocation autocompletion doesn't provide FunctionTypes.
This commit resolves https://github.com/ballerina-platform/ballerina-lang/issues/9423
