+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/completions/resolvers/AbstractItemResolver.java
@@ -67,7 +67,9 @@ public abstract class AbstractItemResolver {
             BSymbol bSymbol = symbolInfo.getScopeEntry() != null ? symbolInfo.getScopeEntry().symbol : null;
             if (!(bSymbol != null && bSymbol.getName().getValue().startsWith("$"))) {
                 if ((bSymbol instanceof BInvokableSymbol
-                        && SymbolKind.FUNCTION.equals(((BInvokableSymbol) bSymbol).kind))
+                        && (((BInvokableSymbol) bSymbol).kind == null &&
+                        SymbolKind.RECORD.equals(((BInvokableSymbol) bSymbol).owner.kind)
+                        || SymbolKind.FUNCTION.equals(((BInvokableSymbol) bSymbol).kind)))
                         || symbolInfo.isIterableOperation()) {
                     completionItem = this.populateBallerinaFunctionCompletionItem(symbolInfo);
                 } else if (!(bSymbol instanceof BInvokableSymbol)
